+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Breaking Change
+* Condensed onChange function props into a single prop that returns an Object for the onsetDate.
+* Removed prop `granularitySelectOnChange`
+* Removed prop `precisionSelectOnChange`
+* Removed prop `onsetDateInputOnChange`
+* New prop `ageUnit`
+* New prop `onsetOnChange`
 
 3.4.0 - (February 26, 2019)
 ----------